Noise Classification Using MATLAB
In this project developed a system that automatically classifies
environment noise in three normally encountered noise environments
Machinery, Street, and Babble. The raw data is collected in Data
collection process in a .wave  format. Prime features are selected from
the raw data for classification. The last block is a classifier. The
focal point of the project is to implement different classifier which
are Neural Network (NN), K-Nearest Neighbour (KNN), Support Vector
Machine (SVM), and Gaussian Mixture Model (GMM).